requirements.txt
flask==3.0.0
pandas==2.1.1
numpy==1.25.2
matplotlib==3.7.2
plotly==5.15.0
python-dotenv==1.0.0 